Move date functions to seperate package
