+++ internal/controllers/device.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/DIMO-Network/rewards-api/internal/services"
+	"github.com/DIMO-Network/rewards-api/internal/date"
 	"github.com/DIMO-Network/rewards-api/models"
 	"github.com/DIMO-Network/shared/db"
 	"github.com/ericlagergren/decimal"
@@ -51,8 +51,8 @@ func (r *DeviceController) GetDevice(c *fiber.Ctx) error {
 		}
 
 		out[i] = DeviceWeek{
-			Start:     services.NumToWeekStart(r.IssuanceWeekID),
-			End:       services.NumToWeekEnd(r.IssuanceWeekID),
+			Start:     date.NumToWeekStart(r.IssuanceWeekID),
+			End:       date.NumToWeekEnd(r.IssuanceWeekID),
 			Tokens:    tokens,
 			VehicleID: r.UserDeviceTokenID.Int(nil),
 		}
