+++ internal/date/date.go
@@ -1,0 +1,34 @@
+// Package date provides functions for working with dates and times in the context of DIMO Rewards.
+package date
+
+import "time"
+
+var (
+	startTime    = time.Date(2022, time.January, 31, 5, 0, 0, 0, time.UTC)
+	weekDuration = 7 * 24 * time.Hour
+)
+
+// GetWeekNum calculates the number of the week in which the given time lies for DIMO point
+// issuance, which at the time of writing starts at 2022-01-31 05:00 UTC. Indexing is
+// zero-based.
+func GetWeekNum(t time.Time) int {
+	sinceStart := t.Sub(startTime)
+	weekNum := int(sinceStart.Truncate(weekDuration) / weekDuration)
+	return weekNum
+}
+
+// GetWeekNumForCron calculates the week number for the current run of the cron job. We expect
+// the job to run every Monday at 05:00 UTC, but due to skew we just round the time.
+func GetWeekNumForCron(t time.Time) int {
+	sinceStart := t.Sub(startTime)
+	weekNum := int(sinceStart.Round(weekDuration) / weekDuration)
+	return weekNum
+}
+
+func NumToWeekStart(n int) time.Time {
+	return startTime.Add(time.Duration(n) * weekDuration)
+}
+
+func NumToWeekEnd(n int) time.Time {
+	return startTime.Add(time.Duration(n+1) * weekDuration)
+}
